refactor: refactor notification hooks to use services directory
